Swap RecipeList and SelectedRecipeList names
